+++ jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  testEnvironment: 'node',
+  testMatch: ['<rootDir>/test/unit/**/*.test.js'],
+  verbose: true
+};